# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flashcards API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/answers": {
            "post": {
                "tags": [
                    "answer"
                ],
                "summary": "Create an answer",
                "operationId": "a1aa30cc6ce16b56a43129192929677d",
                "parameters": [
                    {
                        "name": "flashcardId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string"
                                    },
                                    "explanation": {
                                        "type": "string"
                                    },
                                    "is_correct": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Flashcard not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/answers/{id}": {
            "get": {
                "tags": [
                    "answer"
                ],
                "summary": "Get an answer",
                "operationId": "9fbab7e05c455ae2d2b0f94968eef7e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Answer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "answer"
                ],
                "summary": "Delete an answer",
                "operationId": "e173b1afc518f767162047993b50f321",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "answer"
                ],
                "summary": "Update an answer",
                "operationId": "550bfb18c699b6a03daccdbfce7eed6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string"
                                    },
                                    "explanation": {
                                        "type": "string"
                                    },
                                    "is_correct": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attempt": {
            "get": {
                "tags": [
                    "attempt"
                ],
                "summary": "List attempts",
                "description": "Return all attempts for the current user, paginated",
                "operationId": "17be1018f13f02dc5c6e77cafd33d828",
                "parameters": [
                    {
                        "name": "tags",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attempt/{attempt}": {
            "get": {
                "tags": [
                    "attempt"
                ],
                "summary": "Show an attempt",
                "description": "Return an attempt",
                "operationId": "bd6cf621af769869be0ea6f721f853f0",
                "parameters": [
                    {
                        "name": "attempt",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/tags": {
            "get": {
                "tags": [
                    "tag"
                ],
                "summary": "List tags",
                "operationId": "4d0f4baf9e6fa992c0d85fbf140ebf17",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "tag"
                ],
                "summary": "Create a tag",
                "operationId": "a0a246f29dcb119a2189fd230fc10028",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "colour"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mathematics"
                                    },
                                    "colour": {
                                        "type": "string",
                                        "example": "green"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}